Mechanically move protos and structs for units.cc from defs.h to units.h
